Set aeson version bounds
We will rely on this minimum version to use a simpler import. It wasn't
much trouble to bring in the extra-dep on old resolvers. If it were (or
becomes that way in the future, or we get a user report about it), we
can use CPP to support older aeson -- I'm not against it outright.