Tag SomeException with the underlying exception class
* Tag SomeException with the underlying exception class

The library was calling `exErrorClass` on the `SomeException`. This is
fine, but it always exposes a `SomeException` constructor tag. Since
`SomeException` is an existential type of `Exception`, which has a
superclass of `Typeable`, we can open up the constructor and call
`exErrorClass` on the inner type without ambiguity.

This requires somewhat of an awkward formulation to work properly. The
original `bugsnagExceptionWithParser` call uses `SomeException` via
`parseStringException`, so we must first create a `BugsnagException` and
then edit it with our existential tag. The `SomeException` must be
unwrapped to do this, so we need apply a lambda to do so.

* Remove redundant casters

These casters were in place to coerce `SomeException` types to a list of
standard exceptions. With existential unwrapping of the constructor tag
these castors provide no other utility. Coverage for this is validated
by the `IOException` doc test.