fix(bugsnag): handle ExceptionWithContext
GHC-9.12 / base-4.20 introduces `ExceptionWithContext` as a way to read
backtraces off of functions like `error`. This means `ErrorCall` no
longer includes the information, which broke us.

To address this, we can `fromException` to that type, and then parse the
displaying of it. The format is similar but different to what
`ErrorCall` was, but we can re-use most of `errorCallParser` by
factoring out the stack-frame bits.

To keep things working on older GHC's we can put this behind CPP and
expose our own type and function of the same names, but with no-op
functionality.

Fixes #99.