chore(ci): ensure coverage works
We were doing that only for the `stack.yaml` job, but the setup now has
`stack.yaml` as a symlink to the latest LTS, and _not_ running CI for it
duplicatively. So instead we need to maintain the "default" job as the
latest LTS concretely.

I wonder if `generate-matrix` could put something else in the matrix to
identify where a given resolver sits in the set, that we could key on my
dynamically.